requirements.txt
django
djangorestframework
pymilvus==2.3.4
sentence-transformers
